Completed chapter 6!
Created page to show previous TacoOrders

Information on application properties is interesting, but the book is out of data with it having been based on Spring 2.something. Additionally, the author is using XML in the examples while I am not.

Review information in a different book: Spring Boot Up and Running Mark Heckler