Chap 8 Section 8.2 - sort of completed....
Was only able to get the consent pages to work by commenting out both AuthorizationServerConfig.authorizationServerSecurityFilterChain() and SecurityConfig.securityFilterChain()

Why?!?

Book is not any help and is outdated with the major revisions done to Spring Boot 3 and the various dependencies since book was published. However, even building running the author's source code the toggles on the consent page do not appear.

This book is not very good....